refactor: move MetricReaderFactory to reader module, rename to MetricReaderFactoryImpl (#121)
