+++ src/main.rs
@@ -17,133 +17,12 @@ use clap::Parser;
 use tokio_util::sync::CancellationToken;
 use tracing::{error, info};
 
-use collector::{CollectorEngine, MetricReaderFactory, DEFAULT_SHUTDOWN_TIMEOUT};
-use config::{find_config_file, Cli, Config, Protocol};
+use collector::{CollectorEngine, DEFAULT_SHUTDOWN_TIMEOUT};
+use config::{find_config_file, Cli, Config};
 use internal_metrics::InternalMetrics;
 use logging::{init_logging, LogOutput, LoggingConfig};
 use metrics::MetricStore;
-use reader::modbus::{rtu::ModbusRtuMetricReader, tcp::ModbusTcpMetricReader};
-
-// ── Real metric reader factory ────────────────────────────────────────
-
-struct RealMetricReaderFactory;
-
-impl MetricReaderFactory for RealMetricReaderFactory {
-    fn create(&self, collector: &config::CollectorConfig) -> Result<Box<dyn reader::MetricReader>> {
-        match &collector.protocol {
-            Protocol::ModbusTcp { endpoint } => {
-                let slave_id = collector.slave_id.unwrap_or(1);
-                Ok(Box::new(ModbusTcpMetricReader::new(
-                    endpoint.clone(),
-                    slave_id,
-                )))
-            }
-            Protocol::ModbusRtu {
-                device,
-                bps,
-                data_bits,
-                stop_bits,
-                parity,
-            } => {
-                let slave_id = collector.slave_id.unwrap_or(1);
-                let builder = tokio_serial::new(device, *bps)
-                    .data_bits(match data_bits {
-                        5 => tokio_serial::DataBits::Five,
-                        6 => tokio_serial::DataBits::Six,
-                        7 => tokio_serial::DataBits::Seven,
-                        _ => tokio_serial::DataBits::Eight,
-                    })
-                    .stop_bits(match stop_bits {
-                        2 => tokio_serial::StopBits::Two,
-                        _ => tokio_serial::StopBits::One,
-                    })
-                    .parity(match parity {
-                        config::Parity::None => tokio_serial::Parity::None,
-                        config::Parity::Even => tokio_serial::Parity::Even,
-                        config::Parity::Odd => tokio_serial::Parity::Odd,
-                    });
-                Ok(Box::new(ModbusRtuMetricReader::new(builder, slave_id)))
-            }
-            Protocol::I2c { bus, address } => {
-                #[cfg(target_os = "linux")]
-                let device: Box<dyn reader::i2c::I2cDevice> = {
-                    let mut dev =
-                        reader::i2c::linux_device::LinuxI2cDevice::new(bus.clone(), *address);
-                    dev.open().context("failed to open I2C device")?;
-                    Box::new(dev)
-                };
-                #[cfg(not(target_os = "linux"))]
-                let device: Box<dyn reader::i2c::I2cDevice> = Box::new(reader::i2c::StubI2cDevice);
-
-                let bus_lock = reader::i2c::get_bus_lock(bus);
-                let client =
-                    reader::i2c::I2cMetricReader::new(device, bus.clone(), *address, bus_lock);
-                Ok(Box::new(client))
-            }
-            Protocol::Spi {
-                device,
-                speed_hz,
-                mode,
-                bits_per_word,
-            } => {
-                #[cfg(target_os = "linux")]
-                let spi_device: Box<dyn reader::spi::SpiDevice> = {
-                    let mut dev = reader::spi::linux_device::LinuxSpiDevice::new(
-                        device.clone(),
-                        *speed_hz,
-                        *mode,
-                        *bits_per_word,
-                    );
-                    dev.open().context("failed to open SPI device")?;
-                    Box::new(dev)
-                };
-                #[cfg(not(target_os = "linux"))]
-                let spi_device: Box<dyn reader::spi::SpiDevice> =
-                    Box::new(reader::spi::StubSpiDevice);
-
-                let device_lock = reader::spi::get_device_lock(device);
-                let client =
-                    reader::spi::SpiMetricReader::new(spi_device, device.clone(), device_lock);
-                Ok(Box::new(client))
-            }
-            Protocol::I3c {
-                bus,
-                pid,
-                address,
-                device_class,
-                instance,
-            } => {
-                let address_mode = if let Some(pid_str) = pid {
-                    reader::i3c::AddressMode::Pid(pid_str.clone())
-                } else if let Some(addr) = address {
-                    reader::i3c::AddressMode::Static(*addr)
-                } else {
-                    reader::i3c::AddressMode::DeviceClass {
-                        class: device_class.clone().unwrap(),
-                        instance: instance.unwrap(),
-                    }
-                };
-
-                #[cfg(target_os = "linux")]
-                let device: Box<dyn reader::i3c::I3cDevice> = {
-                    let mut dev = reader::i3c::linux_device::LinuxI3cDevice::new(bus.clone());
-                    dev.open().context("failed to open I3C device")?;
-                    Box::new(dev)
-                };
-                #[cfg(not(target_os = "linux"))]
-                let device: Box<dyn reader::i3c::I3cDevice> = Box::new(reader::i3c::StubI3cDevice);
-
-                let client = reader::i3c::I3cMetricReader::new(device, bus.clone(), address_mode);
-                let bus_lock = reader::i3c::get_bus_lock(bus);
-                let handle = reader::i3c::I3cMetricReaderHandle::new(
-                    std::sync::Arc::new(tokio::sync::Mutex::new(client)),
-                    bus_lock,
-                );
-                Ok(Box::new(handle))
-            }
-        }
-    }
-}
+use reader::MetricReaderFactoryImpl;
 
 // ── Config → logging mapping ──────────────────────────────────────────
 
@@ -223,7 +102,7 @@ async fn main() -> Result<()> {
     // 5. Spawn collector tasks
     let global_labels: BTreeMap<String, String> =
         config.global_labels.clone().into_iter().collect();
-    let factory = RealMetricReaderFactory;
+    let factory = MetricReaderFactoryImpl;
     let engine = CollectorEngine::spawn(
         config.collectors.clone(),
         store.clone(),
